config: move enterprise_edition to schema module
The enterprise_edition() schema helper -- which marks a schema node as
available only in Tarantool Enterprise Edition -- was file-local to
instance_config. Move it to the experimental.config.utils.schema module
and export it, so other config schemas (for example, the recovery point
backends) can reuse it instead of re-implementing the edition check.

instance_config now uses the exported helper; behavior is unchanged.

Needed for #12865

NO_DOC=refactoring
NO_CHANGELOG=internal
NO_TEST=refactoring
