+++ src/box/lua/config/utils/schema.lua
@@ -18,6 +18,8 @@
 local fun = require('fun')
 local json = require('json')
 local units = require('internal.config.utils.units')
+local tarantool = require('tarantool')
+local funcutils = require('internal.config.utils.funcutils')
 
 local json_noexc = json.new()
 json_noexc.cfg({encode_use_tostring = true})
@@ -2954,6 +2956,61 @@ end
 
 -- }}} schema.fromenv()
 
+-- {{{ Schema node annotators
+
+local function enterprise_edition_validate(data, w)
+    -- OK if we're on Tarantool EE.
+    if tarantool.package == 'Tarantool Enterprise' then
+        return
+    end
+
+    assert(tarantool.package == 'Tarantool')
+
+    -- OK, if the value is nil or box.NULL.
+    if data == nil then
+        return
+    end
+
+    -- NB: Let's fail the validation for an empty table, because
+    -- otherwise we will get a less descriptive error from a lower
+    -- level API. For example, box.cfg({wal_ext = {}}) on Tarantool
+    -- Community Edition says the following:
+    --
+    -- > Incorrect value for option 'wal_ext': unexpected option
+
+    w.error('This configuration parameter is available only in Tarantool ' ..
+        'Enterprise Edition')
+end
+
+local function enterprise_edition_apply_default_if(_data, _w)
+    return tarantool.package == 'Tarantool Enterprise'
+end
+
+-- Mark a schema node as available only in Tarantool Enterprise Edition:
+-- configuring it on Community Edition fails validation with a clear message
+-- (instead of a less descriptive lower-level error), and its default is applied
+-- only on Enterprise Edition.
+local function enterprise_edition(schema_node)
+    schema_node.enterprise_edition = true
+
+    -- Perform a domain specific validation first and only then
+    -- check, whether the option is to be used with Tarantool
+    -- Enterprise Edition.
+    --
+    -- This order is consistent with data type validation, which
+    -- is also performed before the EE check.
+    schema_node.validate = funcutils.chain2(
+        schema_node.validate,
+        enterprise_edition_validate)
+    schema_node.apply_default_if = funcutils.chain2(
+        schema_node.apply_default_if,
+        enterprise_edition_apply_default_if)
+
+    return schema_node
+end
+
+-- }}} Schema node annotators
+
 return {
     -- Schema node constructors.
     scalar = scalar,
@@ -2975,4 +3032,7 @@ return {
 
     -- Parse data from an environment variable.
     fromenv = fromenv,
+
+    -- Schema node annotators.
+    _enterprise_edition = enterprise_edition,
 }
